added a query to get answers with photos
