+++ sequelize/photos.js
@@ -1,18 +1,18 @@
-const db = require('./connections.js');
+// const db = require('./connections.js');
 
-let Photos = db.define('photos', {
-  id: {
-    type: Sequelize.INTEGER,
-    allowNull: false,
-    autoIncrement: true,
-    primaryKey: true
-  },
-  answer_id: {
-    type: Sequelize.INTEGER,
-    allowNull: false
-  },
-  url: {
-    type: Sequelize.STRING,
-    allowNull: false
-  }
-});
+// let Photos = db.define('photos', {
+//   id: {
+//     type: Sequelize.INTEGER,
+//     allowNull: false,
+//     autoIncrement: true,
+//     primaryKey: true
+//   },
+//   answer_id: {
+//     type: Sequelize.INTEGER,
+//     allowNull: false
+//   },
+//   url: {
+//     type: Sequelize.STRING,
+//     allowNull: false
+//   }
+// });
